+++ android/app/src/main/kotlin/com/example/chamber/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.chamber
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}