+++ lib/l10n/app_localizations.dart
@@ -5,29 +5,45 @@ import 'package:flutter/services.dart';
 
 class AppLocalizations {
   final Locale locale;
-  late Map<String, String> _localizedStrings;
+  Map<String, String>? _localizedStrings;
 
   AppLocalizations(this.locale);
 
+  // Helper method to load the JSON file from the assets
+  Future<String> _loadJsonFromAssets(String path) async {
+    try {
+      return await rootBundle.loadString(path);
+    } catch (e) {
+      // If an exception occurs, log the error and return a default empty JSON object
+      debugPrint('Could not load the json file: $e');
+      return '{}'; // Return an empty JSON object to prevent further exceptions
+    }
+  }
+
   Future<bool> load() async {
-    String jsonPath;
+    String jsonPath = 'assets/l10n/${locale.languageCode}.json';
     if (locale.languageCode == 'pt' && locale.countryCode == 'BR') {
-      jsonPath = 'l10n/pt.json'; // Adjust the path if needed
-    } else {
-      jsonPath = 'l10n/${locale.languageCode}.json';
+      jsonPath = 'assets/l10n/pt_BR.json'; // Adjust the path if needed
     }
 
-    String jsonString = await rootBundle.loadString(jsonPath);
+    // Use the helper method to load the JSON file safely
+    String jsonString = await _loadJsonFromAssets(jsonPath);
     Map<String, dynamic> jsonMap = json.decode(jsonString);
+
+    // Convert the dynamic values to String
     _localizedStrings = jsonMap.map((key, value) => MapEntry(key, value.toString()));
     return true;
   }
 
   String translate(String key) {
-    return _localizedStrings[key] ?? "";
+    return _localizedStrings?[key] ?? "";
   }
 
   static AppLocalizations of(BuildContext context) {
-    return Localizations.of<AppLocalizations>(context, AppLocalizations)!;
+    // Attempt to get the instance of AppLocalizations
+    AppLocalizations? appLocale = Localizations.of<AppLocalizations>(context, AppLocalizations);
+
+    // Return the found instance or a default instance for 'en_US' if none is found
+    return appLocale ?? AppLocalizations(Locale('en', 'US'));
   }
 }