+++ CHANGELOG.md
@@ -9,3 +9,9 @@
 - Added localizations. Languages supported: English (en), Spanish (es), Brazilian Portuguese (pt-BR), and Chinese (zh). Open an issue if you would like to request additional languages.
 - Introduced ChamberDialog, an enhanced widget for viewing logs, featuring search functionality and other features.
 - Implemented a configuration object to customize maxLogs, date formatting and user authorization.
+
+## 1.1.1
+
+- Fixes to localization and added full app example.
+- Enabling android to test example.
+- Few changes to dialog widget.